requirements.txt
langchain-core
python-dotenv
streamlit
langchain-groq
langchain_community
